third_party/WebKit/LayoutTests/http/tests/inspector/bindings/navigator-frame-navigate-expected.txt

Verify that navigator is updated as frame gets navigated.



Running: dumpInitialNavigator
top
  127.0.0.1:8000
    inspector
      bindings
        navigator-frame-navigate.html
        bindings-test.js
      inspector-test.js

Running: attachFrame
top
  127.0.0.1:8000
    inspector
      bindings
        navigator-frame-navigate.html
        bindings-test.js
      inspector-test.js
  frame (magic-frame.html)
    127.0.0.1:8000
      inspector/bindings/resources
        magic-frame.html
        magic-script.js
        magic-style.css

Running: navigateFrame
top
  127.0.0.1:8000
    inspector
      bindings
        navigator-frame-navigate.html
        bindings-test.js
      inspector-test.js
  frame (empty-frame.html)
    127.0.0.1:8000
      inspector/bindings/resource
        empty-frame.html
